third_party/blink/web_tests/wpt_internal/webaudio/setsinkid-constructor-onerror.https-expected.txt

This is a testharness.js-based test.
[FAIL] AudioContext constructor with an invalid sinkId
  assert_equals: Sink ID expected "" but got "an-invalid-sink-id"
Harness: the test ran to completion.
